requirements.txt
Flask==3.0.3
google-generativeai
python-dotenv
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
torch